pyproject.toml
[tool.poetry]
name = "untube"
version = "0.1.0"
description = ""
authors = ["Mohammed Khan <abukhan363@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
# psycopg2 = "^2.9.6"
django-crispy-forms = "^2.0"
bleach = "^6.0.0"
humanize = "^4.6.0"
google-auth = "^2.18.1"
google-api-python-client = "^2.86.0"
django-allauth ="^0.47.0"  # "^0.54.0"
whitenoise = "^6.4.0"
django-split-settings = "^1.2.0"
pyyaml = "^6.0"
django-import-export = "^3.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
